Use a dedicated scenario file for agentless end-to-end
