+++ tests/test_the_test/test_mock_ffe_agentless_backend.py
@@ -7,6 +7,7 @@ import requests
 import pytest
 
 from utils import scenarios
+from utils._context._scenarios import agentless_endtoend as agentless_endtoend_scenarios
 from utils._context._scenarios import endtoend as endtoend_scenarios
 from utils.docker_fixtures._core import HOST_GATEWAY_EXTRA_HOSTS, extra_hosts_for_environment
 from utils.mocked_backend.ffe import (
@@ -16,7 +17,7 @@ from utils.mocked_backend.ffe import (
     MockFFEAgentlessBackendServer,
     UFC_RESPONSE_TYPE,
 )
-from utils._context._scenarios.endtoend import FeatureFlaggingAgentlessEndToEndScenario
+from utils._context._scenarios.agentless_endtoend import FeatureFlaggingAgentlessEndToEndScenario
 
 
 @scenarios.test_the_test
@@ -135,7 +136,7 @@ def test_agentless_end_to_end_scenario_closes_backend_when_startup_fails(
     def create_backend() -> MagicMock:
         return backend
 
-    monkeypatch.setattr(endtoend_scenarios, "MockFFEAgentlessBackendServer", create_backend)
+    monkeypatch.setattr(agentless_endtoend_scenarios, "MockFFEAgentlessBackendServer", create_backend)
 
     with pytest.raises(RuntimeError, match="reset failed"):
         scenario.configure(MagicMock(spec=pytest.Config))
